LayoutTests/http/tests/websocket/tests/hybi/inspector/server-close-expected.txt

Tests WebSockets connection closed by the server.


== Running test suite: WebSocket.ServerClose
-- Running test case: WebSocket.ServerClose
PASS: WebSocketResource should be created.
PASS: WebSocketResource should be a WebSocket type.
PASS: WebSocketResource should be in a Connecting state.
PASS: WebSocketResource previous state should be Connecting.
PASS: WebSocketResource state should change to Open.
PASS: WebSocketResource should be in a Open state.
PASS: WebSocketResource previous state should be Open.
PASS: WebSocketResource state should change to Closed.
PASS: WebSocketResource should be in a Closed state.
